+++ Plugin/Thunderstore/CHANGELOG.md
@@ -1,4 +1,4 @@
-## v0.0.4
+## v0.1.0
 - Added new item, SnowGlobe.
 - Added configs for SnowGlobe.
 - Added a loooooot of configs.
@@ -9,6 +9,11 @@
 - Improved meteor spawning location.
 - Added new shop item, hoverboard.
 - Added config and chance for critical hit from code rebirth weapons.
+- Added tornado.
+- Updated readme.
+- Updated all icons.
+- Updated meteorite Crystals.
+- Improved how wallet works, it's a player upgrade now.
 
 ## v0.0.3
 - Added moon blacklist config for the weather.
@@ -33,6 +38,4 @@
 - Initial release
 
 ## TODO:
-- Add Tornado
-- Add ScrapMaster
 - Updated Icons for items